Separate existing tests into different files.
